pyproject.toml
[tool.poetry]
name = "shylock"
version = "1.2.0"
description = "Distributed locks in Python, similar to https://github.com/vaidik/sherlock - also with asyncio support"
authors = ["Janne Enberg <janne.enberg@lietu.net>"]
license = "BSD-3-Clause"
readme = "README.rst"
homepage = "https://github.com/lietu/shylock/"
repository = "https://github.com/lietu/shylock/"
documentation = "https://github.com/lietu/shylock/"
keywords = ["distributed", "locking", "lock", "asyncio"]

packages = [
    { include = "shylock" },
]

[tool.poetry.dependencies]
python = "^3.7"
aioarangodb = { version = "^0.1.2", optional = true }
motor = { version = "^2.0.0", optional = true }
pymongo = { version = "^3.10.1", optional = true }
python-arango = { version = "^5.4.0", optional = true }

[tool.poetry.dev-dependencies]
pytest = "^7.2.1"

[tool.poetry.extras]
aioarangodb = ["aioarangodb"]
motor = ["motor"]
pymongo = ["pymongo"]
python-arango = ["python-arango"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
target_version = ["py36"]

[tool.isort]
profile = "black"
